wip
git-svn-id: https://www.dkfz.de/svn/B040/FCSTOOLS/trunk/QuickFit3@3455 0d8be23c-4016-4ca5-a69c-3788035ecc6c

+++ plugins/cam_andor/cam_andor.cpp
@@ -773,7 +773,7 @@ QVariant QFExtensionCameraAndor::getCameraSetting(QSettings& settings, QFExtensi
 
 QVariant QFExtensionCameraAndor::getCameraCurrentSetting(int camera, QFExtensionCamera::CameraSetting which)
 {
-    if (camera<0 || camera>=getCameraCount()) return QVariant();
+    if (camera<0 || camera>=(long long)getCameraCount()) return QVariant();
 
     if (which==QFExtensionCamera::CamSetExposureTime) return camInfos[camera].expoTime;
     if (which==QFExtensionCamera::CamSetNumberFrames) return camInfos[camera].numKins;
